requirements.txt
pandas
langdetect
emoji
beautifulsoup4
tqdm
supabase
python-dateutil
app-store-web-scraper
python-dotenv
